Extract synchronization classes
